backend/requirements.txt
attrs==19.3.0
blis==0.2.4
boto==2.49.0
boto3==1.9.146
botocore==1.12.253
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cymem==2.0.3
docutils==0.15.2
en-core-web-sm==2.0.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
gensim==3.7.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.0.2
MarkupSafe==1.1.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.2
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
s3transfer==0.2.1
scipy==1.4.1
six==1.14.0
smart-open==1.11.1
spacy==2.1.4
SQLAlchemy==1.3.16
srsly==1.0.2
thinc==7.0.8
tqdm==4.45.0
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
